requirements.txt
certifi==2025.3.23
charset-normalizer==3.3.2
idna==3.7
msal==1.28.0
python-dotenv==1.0.1
requests==2.32.2
urllib3==2.2.1
